Now using DCLowerLimit for kinetics
Used to drip in just the dissolved IC moles into the solution, but now
we make the kinetic phases part of the thermodynamic system and use the
DCLowerLimit to force them to dissolve at the rate we want.

This approach seems to work for ParrotKilloh kinetics but it crashes
when applied to StandardKineticModel.